Added confirm delete popup to all delete buttons

+++ src/main/java/com/example/c482project/AddProductController.java
@@ -1,6 +1,5 @@
 package com.example.c482project;
 
-import javafx.collections.ObservableList;
 import javafx.collections.transformation.FilteredList;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
@@ -16,8 +15,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.function.Predicate;
 
-import static com.example.c482project.c482Project.inventory;
-import static com.example.c482project.c482Project.productID;
+import static com.example.c482project.c482Project.*;
 
 //TODO•   The user should not delete a product that has a part associated with it.
 
@@ -72,8 +70,10 @@ public class AddProductController {
     }
 
     public void removeAssociatedPartBtnClick() {
-        newProduct.deleteAssociatedPart(associatedPartTableView.getFocusModel().getFocusedItem());
-
+        confirmPopup.showAndWait();
+        if (confirmDelete) {
+            newProduct.deleteAssociatedPart(associatedPartTableView.getFocusModel().getFocusedItem());
+        }
     }
     public void saveBtnClick() throws IOException {
         InputVerification.resetErrorText(errorTextList);
@@ -123,7 +123,7 @@ public class AddProductController {
 
     public void initialize() {
         errorTextList.addAll(Arrays.asList(nameErrorText,invErrorText,priceErrorText,minMaxErrorText));
-
+        idTextField.setText(String.valueOf(productID));
 
         newProduct = new Product(productID,null,-1,-1,-1,-1);
         // Init top table
